refactor: Split database types based on services
